refactor(torghut): remove duplicate proof target fetcher
